pg_to_redis_listener/requirements.txt
psycopg2-binary
redis